refactor: move middlewares & add initial file caching middleware
