--- src/middlewares/index.ts
@@ -1,1 +1,0 @@
-export * from "./cache-middleware";
